Stage 9: split RenderDebugGUISystem into Editor/Panels files
RenderDebugGUISystem.cpp had grown to a 1179-LOC editor god-UI owning the
menu bar, playback toolbar, every editor window, the build/save-layout
modals, SDL file-dialog callbacks, the ImGui font/style backend, and the
per-frame frame begin/end — all behind one Render().

Split it into focused units under src/Editor/Panels/ (namespace
octarine::editor::panels):

  EditorPanelHost      dockspace + per-frame dispatch + ApplyPendingLayout /
                       DrawProjectSelectorIfNeeded / ProcessFontRebuild hooks
  EditorToolbarPanel   main menu bar + playback toolbar
  EditorBuildDialog    Export Build + Save Layout modals
  EditorImGuiBackend   font atlas rebuild + editor style (lifted from header)
  EditorDialogService  SDL file/folder dialog callbacks
  EditorSceneWindow    game-texture composite + viewport sync
  EditorHierarchyPanel entity tree + component inspector host
  EditorAssetBrowserPanel / EditorConsolePanel / EditorOutputPanels /
  EditorProjectPanels / EditorSettingsPanels / EditorSigningPanel

RenderDebugGUISystem now keeps only the IMGUI-shared frame begin/end and the
game overlays (onDebugGUI script hook, FPS plot, entity count) and delegates
the editor chrome to the host. Its .cpp drops 1179 -> 104 LOC and the header
sheds its EditorPersistence / Roboto / imgui includes. Game.cpp's font-init
calls repoint to octarine::editor::panels::{RebuildEditorFont,ApplyEditorStyle}.

The new panel TUs are editor-only (OCTARINE_WITH_EDITOR) in the CMake source
list; RenderDebugGUISystem.cpp stays on the IMGUI list so the player-debug
build still gets the game overlays without the editor.

No behavior change. Verified: editor-debug, editor-release (-Werror),
player-debug (gating: 0 panel objects, 1 system object), and player-release
all configure + build clean.
